Git rid of "undefine all vars" stuff, since I'm going to use an alternative method to verify that we have a valid model DAG.
Implemented case where the LHS of a node draw (when drawing from the full joint distribution) is an array slice.

Added a reader macro: now you can write #e<sexpr> instead of (sexpr->expr '<sexpr>).

Added support for :shape keyword to quantifier expressions; this is needed for things like q@sum, where the shape of array to return is not known if the quantification range is empty.